+++ src/audio/constant.ts
@@ -1,0 +1,4 @@
+export const MAX_RAND = 8192;
+export const MIN_FREQ = 800;
+export const GRAD = 0.02444;
+export const MAX_FREQ = GRAD * MAX_RAND + MIN_FREQ;
